Update for Data-Driven Methods (04-04-25)
This update contains the initial trial for accessing the openAI API and using the GPT 4o-mini model to generate a messy mesh text file output, run it through a parsing algorithm to clean it and structure it properly into Vertices and Faces. Then the resulting text file is run through an algorithm that uses the ifcopenshell API to generate an ifc model from the obj mesh. This can then be tried to open in Revit (once it is debugged and works properly).